+++ src/main/java/com/test/rag/model/ScoredChunk.java
@@ -1,6 +1,8 @@
 package com.test.rag.model;
 
+import java.math.BigDecimal;
+
 public record ScoredChunk(
         DocumentChunk chunk,
-        double similarityScore
+        BigDecimal similarityScore
 ) {}
